Small tweaks to processing. URL processing of upload is not working for new Website.

+++ helper_functions/clearScreenshotsWorkingDir.js
@@ -3,12 +3,15 @@ import fs from "fs";
 export default (dir) => {
     // const parentDir = dir.slice(0, dir.indexOf("/"));
     const fullPathParentDir = `app/projects/${dir}/screenshots`;
-    const parentDirFolders = fs.readdirSync(fullPathParentDir);
+    let parentDirFolders = [];
 
-    parentDirFolders.map((path) => {
-        fs.rmSync(`${fullPathParentDir}/${path}`, {
-            recursive: true,
-            force: true,
+    if (fs.existsSync(fullPathParentDir)) {
+        parentDirFolders = fs.readdirSync(fullPathParentDir);
+        parentDirFolders.map((path) => {
+            fs.rmSync(`${fullPathParentDir}/${path}`, {
+                recursive: true,
+                force: true,
+            });
         });
-    });
+    } else return;
 };
